refactor: feature module layout — co-located .h, .cpp, tests per module (ADR-018)
Moved from flat src/include/test dirs to feature folders:
  src/config/    src/json/      src/ollama/    src/repl/
  src/command/   src/annotation/ src/exec/     src/tui/

Each module contains its .h, .cpp, _test.cpp, and _it.cpp.
Include path changed from include/ to src/.
Removed empty include/ and test/ dirs (except shell scripts).

Closes #35
